fix(cvIO): avoid conversion in matrix helper, use opencv mat44d instead

+++ SrcLib/io/cvIO/include/cvIO/Matrix.hpp
@@ -15,7 +15,7 @@
 namespace cvIO
 {
 /**
- * @brief The Matrix class
+ * @brief The Matrix helper class contains static method to convert ::fwData::TransformationMatrix3D to opencv Mat.
  */
 class CVIO_CLASS_API Matrix
 {
@@ -23,7 +23,7 @@ public:
 
     /**
      * @brief copyFromCv: copies values from OpenCV structure to fw4spl data.
-     * @param _src [::cv::Matx44f]: Opencv 4x4 float matrix.
+     * @param _src [::cv::Matx44f]: OpenCV 4x4 float matrix.
      * @param _dst [::fwData::TransformationMatrix3D::sptr]: TranformationMatrix (need to be initialized).
      */
     CVIO_API static void copyFromCv( const ::cv::Matx44f& _src, ::fwData::TransformationMatrix3D::sptr& _dst );
@@ -31,9 +31,9 @@ public:
     /**
      * @brief copyToCv: copies values from fw4spl TransformationMatrix3D to OpenCV structure.
      * @param _src [::fwData::TransformationMatrix3D::csptr]: fw4spl matrix to be copied.
-     * @param _dst [::cv::Matx44f]: OpenCV 4x4 float matrix.
+     * @param _dst [::cv::Matx44d]: OpenCV 4x4 double matrix.
      */
-    CVIO_API static void copyToCv(const ::fwData::TransformationMatrix3D::csptr& _src, ::cv::Matx44f& _dst);
+    CVIO_API static void copyToCv(const ::fwData::TransformationMatrix3D::csptr& _src, ::cv::Matx44d& _dst);
 };
 
 } //namespace cvIO
